Refactor: widen TensorMap heap_size and related fields to uint64_t
On 64-bit devices, TensorMap's heap_size was int32_t, capping the
addressable on-device heap at 2 GB. Widen it—and all coupled size,
capacity, and offset fields—to uint64_t.

- Widen core size fields in TensorMap-related structs to uint64_t
- Fix int32_t/uint64_t and size_t/uint64_t mixed-type comparisons
- Add explicit casts at API boundaries to suppress macOS CI warnings
- Reorder PTO2SharedMemoryHeader to eliminate struct padding risk
- Propagate uint64_t to aicpu_build_graph and host_build_graph
- Use PRIx64 for portable uint64_t formatting
- Fix PTO2 task count read to use struct member access

+++ src/runtime/aicpu_build_graph/runtime/runtime.cpp
@@ -63,7 +63,7 @@ bool Runtime::try_set_aicpu_orch_so(const void* data, size_t size) {
         return false;
     }
     memcpy(aicpu_orch_so_storage, data, size);
-    aicpu_orch_so_size = static_cast<uint32_t>(size);
+    aicpu_orch_so_size = size;
     return true;
 }
 
@@ -71,7 +71,7 @@ void Runtime::set_aicpu_orch_so(const void* data, size_t size) { (void)try_set_a
 
 const void* Runtime::get_aicpu_orch_so_data() const { return aicpu_orch_so_size > 0 ? aicpu_orch_so_storage : nullptr; }
 
-size_t Runtime::get_aicpu_orch_so_size() const { return static_cast<size_t>(aicpu_orch_so_size); }
+uint64_t Runtime::get_aicpu_orch_so_size() const { return aicpu_orch_so_size; }
 
 // =============================================================================
 // Task Management
